+++ CHANGELOG.md
@@ -12,6 +12,8 @@ element are combined.
 - Gallery has new callbacks `beforeImage` and `afterImage`.
 - Gallery defaults are accessible with $.featherlightGallery.defaults
 - Callbacks like afterOpen can be set using data-featherlight-after-open
+- Gallery inherits Featherlight's prototype and class methods.
+- Gallery can be called entirely from javascript
 
 ### Breaking changes
 - `$.featherlight` now always trigger a new lightbox. Use `$.fn.featherlight` for attaching events to elements.